requirements.txt
fastapi>=0.115.0
uvicorn[standard]
sse-starlette
claude-agent-sdk
chromadb
httpx
jinja2
python-multipart
python-dotenv
markdown
